Corrected pom.xml build ant unit test, and added build.xml for ant build of executable FixMessageGenerator
git-svn-id: http://to-fix.googlecode.com/svn/trunk@6 86be3960-3010-72e2-a4d8-f94c9a4f7a06

+++ src/test/java/org/tomac/protocol/fix/TestFixContainers.java
@@ -216,7 +216,7 @@ public class TestFixContainers {
 
 			assertNotNull(inMsg);
 			assertTrue(err.hasError());
-			assertEquals(err.sessionRejectReason, FixMessageInfo.SessionRejectReason.REQUIRED_TAG_MISSING);
+			assertEquals(err.sessionRejectReason, (int)FixMessageInfo.SessionRejectReason.REQUIRED_TAG_MISSING);
 
 		}
 	}
